moved tests for SampledData
